--- app/src/main/java/hu/elte/sbzbxr/phoneconnect/ui/ScreenCaptureCallbacks.java
@@ -1,3 +1,0 @@
-package hu.elte.sbzbxr.phoneconnect.ui;
-
-public interface ScreenCaptureCallbacks { }
